+++ src/RadixDlt.NetworkGateway.GatewayApiSdk/Model/StateEntityNonFungibleResourceVaultsPageOptIns.cs
@@ -1,0 +1,6 @@
+﻿namespace RadixDlt.NetworkGateway.GatewayApiSdk.Model;
+
+public partial class StateEntityNonFungibleResourceVaultsPageOptIns
+{
+    public static StateEntityNonFungibleResourceVaultsPageOptIns Default => new ();
+}